pyproject.toml
[project]
name = "Hybrid-AgenticRAG-Qdrant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastembed>=0.7.4",
    "ipykernel>=7.1.0",
    "llama-index>=0.14.12",
    "llama-index-vector-stores-qdrant>=0.9.0",
    "pydantic-ai-slim[openai,web]>=1.39.0",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
    "typer>=0.21.0",
    "uvicorn>=0.40.0",
]
